refactor: refatored code
