+++ src/ApiCalls/index.js
@@ -1,3 +1,4 @@
 import { getData, getCategoryData } from "./getMethodRequest";
+import {loginHandler, signInHandler} from "./postMethodRequest";
 
-export { getData, getCategoryData}
+export { getData, getCategoryData, loginHandler, signInHandler}
