+++ src/App.js
@@ -1,10 +1,18 @@
+import { Route, Routes } from "react-router-dom";
 import "./App.css";
-import {Home} from "./pages/index"
+import {Home, Login, PrivateRoute, SignUp, UserProfile} from "./pages/index"
 
 function App() {
   return (
     <div className="App">
-      <Home />
+      <Routes>
+        <Route path="/" element={<Home />}/>
+        <Route path="/login" element={<Login />} />
+        <Route path="/signup" element={<SignUp />} />
+        <Route path="/user" element={<PrivateRoute />}>
+          <Route path="/user/user-profile" element={<UserProfile />} />
+        </Route>
+      </Routes>
     </div>
   );
 }
