+++ src/Context/index.js
@@ -1,4 +1,5 @@
-import { useData } from "./data.context";
-import { useFilter } from "./filter.context";
+import { useData, DataProvider } from "./data.context";
+import { useFilter, FilterProvider } from "./filter.context";
+import { useAuth , AuthProvider} from "./auth.context";
 
-export {useData, useFilter};
+export {useData, useFilter, useAuth, AuthProvider, DataProvider, FilterProvider};
