+++ src/index.js
@@ -3,19 +3,23 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
-import { DataProvider } from "./Context/data.context";
-import { FilterProvider } from "./Context/filter.context";
+import { BrowserRouter as Router } from "react-router-dom";
+import { AuthProvider, DataProvider, FilterProvider } from "./Context";
 
 // Call make Server
 makeServer();
 
 ReactDOM.render(
   <React.StrictMode>
-    <DataProvider>
-      <FilterProvider>
-        <App />
-      </FilterProvider>
-    </DataProvider>
+    <Router>
+      <AuthProvider>
+        <DataProvider>
+          <FilterProvider>
+            <App />
+          </FilterProvider>
+        </DataProvider>
+      </AuthProvider>
+    </Router>
   </React.StrictMode>,
   document.getElementById("root")
 );
