+++ src/pages/Authentication/index/index.js
@@ -1,0 +1,5 @@
+import { Login } from "../Login/Login";
+import { SignUp } from "../SignUp/SignUp";
+import { PrivateRoute } from "../PrivateRoute";
+
+export {Login, SignUp, PrivateRoute};
