+++ src/pages/index/index.js
@@ -1,3 +1,5 @@
 import { Home } from "../Home/Home";
+import {Login, SignUp, PrivateRoute} from "../Authentication/index";
+import { UserProfile } from "../UserProfile/UserProfile";
 
-export {Home}
+export {Home, Login, SignUp, UserProfile, PrivateRoute};
